Factor out Select component
